refactor: point data enum
