+++ services/API-service/src/api/point-data/point-data.entity.ts
@@ -1,18 +1,7 @@
 import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
 
 import { DynamicPointDataEntity } from './dynamic-point-data.entity';
-
-export enum PointDataEnum {
-  evacuationCenters = 'evacuation_centers',
-  dams = 'dams',
-  healthSites = 'health_sites',
-  redCrossBranches = 'red_cross_branches',
-  communityNotifications = 'community_notifications',
-  schools = 'schools',
-  waterpointsInternal = 'waterpoints_internal',
-  gauges = 'gauges',
-  glofasStations = 'glofas_stations',
-}
+import { PointDataEnum } from './point-data.enum';
 
 @Entity('point-data')
 export class PointDataEntity {
