+++ services/API-service/src/api/point-data/point-data.enum.ts
@@ -1,0 +1,11 @@
+export enum PointDataEnum {
+  evacuationCenters = 'evacuation_centers',
+  dams = 'dams',
+  healthSites = 'health_sites',
+  redCrossBranches = 'red_cross_branches',
+  communityNotifications = 'community_notifications',
+  schools = 'schools',
+  waterpointsInternal = 'waterpoints_internal',
+  gauges = 'gauges',
+  glofasStations = 'glofas_stations',
+}
